move AnimationPercentage into com/wynntils/utils/render/type
